+++ .gitignore
@@ -1,7 +1,7 @@
 ### IntelliJ IDEA ###
 out/
 !**/src/main/**/out/
-!src/test/**/out/
+!src/graph/**/out/
 
 ### Eclipse ###
 .apt_generated
@@ -13,7 +13,7 @@ out/
 .sts4-cache
 bin/
 !**/src/main/**/bin/
-!src/test/**/bin/
+!src/graph/**/bin/
 
 ### NetBeans ###
 /nbproject/private/
