+++ src/configs/Config.java
@@ -1,4 +1,4 @@
-package test;
+package configs;
 
 public interface Config {
     void create();
