+++ src/configs/GenericConfig.java
@@ -1,4 +1,4 @@
-package test;
+package configs;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
