+++ src/configs/Node.java
@@ -1,4 +1,6 @@
-package test;
+package configs;
+
+import graph.Message;
 
 import java.util.ArrayList;
 import java.util.List;
