+++ src/graph/App4.java
@@ -1,5 +1,7 @@
-package test;
+package graph;
 
+
+import configs.GenericConfig;
 
 public class App4 {
     public static void main(String[] args) {
