+++ src/graph/IncAgent.java
@@ -1,29 +1,25 @@
-package test;
+package graph;
 
-public class PlusAgent implements Agent {
+import configs.Agent;
+
+public class IncAgent implements Agent {
 
     private final String[] subscribedTopics;
     private final String[] publishedTopics;
     private final TopicManagerSingleton.TopicManager topicManager;
-    private Double firstOperand;
-    private Double secondOperand;
 
-    public PlusAgent(String[] subscribedTopics, String[] publishedTopics) {
+    public IncAgent(String[] subscribedTopics, String[] publishedTopics) {
         this.subscribedTopics = subscribedTopics;
         this.publishedTopics = publishedTopics;
         this.topicManager = TopicManagerSingleton.get();
 
         initializeSubscriptions();
         initializePublications();
-        resetOperands();
     }
 
     private void initializeSubscriptions() {
         if (subscribedTopics.length > 0) {
             topicManager.getTopic(subscribedTopics[0]).subscribe(this);
-        }
-        if (subscribedTopics.length > 1) {
-            topicManager.getTopic(subscribedTopics[1]).subscribe(this);
         }
     }
 
@@ -33,11 +29,6 @@ public class PlusAgent implements Agent {
         }
     }
 
-    private void resetOperands() {
-        this.firstOperand = 0.0;
-        this.secondOperand = 0.0;
-    }
-
     @Override
     public String getName() {
         return getClass().getSimpleName();
@@ -45,43 +36,27 @@ public class PlusAgent implements Agent {
 
     @Override
     public void reset() {
-        resetOperands();
+        // No state to reset in this agent
     }
 
     @Override
     public void callback(String topic, Message msg) {
-        double messageValue = msg.asDouble;
-
-        if (Double.isNaN(messageValue)) {
+        if (Double.isNaN(msg.asDouble)) {
             return;
         }
 
-        updateOperand(topic, messageValue);
-
-        if (areBothOperandsSet() && hasPublishedTopics()) {
-            publishResult();
+        if (hasPublishedTopics()) {
+            publishIncrementedValue(msg.asDouble);
         }
-    }
-
-    private void updateOperand(String topic, double messageValue) {
-        if (subscribedTopics.length > 0 && topic.equals(subscribedTopics[0])) {
-            firstOperand = messageValue;
-        } else if (subscribedTopics.length > 1 && topic.equals(subscribedTopics[1])) {
-            secondOperand = messageValue;
-        }
-    }
-
-    private boolean areBothOperandsSet() {
-        return firstOperand != null && secondOperand != null;
     }
 
     private boolean hasPublishedTopics() {
         return publishedTopics.length > 0;
     }
 
-    private void publishResult() {
-        double result = firstOperand + secondOperand;
-        topicManager.getTopic(publishedTopics[0]).publish(new Message(result));
+    private void publishIncrementedValue(double originalValue) {
+        double incrementedValue = originalValue + 1.0;
+        topicManager.getTopic(publishedTopics[0]).publish(new Message(incrementedValue));
     }
 
     @Override
@@ -93,9 +68,6 @@ public class PlusAgent implements Agent {
     private void unsubscribeFromTopics() {
         if (subscribedTopics.length > 0) {
             topicManager.getTopic(subscribedTopics[0]).unsubscribe(this);
-        }
-        if (subscribedTopics.length > 1) {
-            topicManager.getTopic(subscribedTopics[1]).unsubscribe(this);
         }
     }
