+++ src/graph/Message.java
@@ -1,4 +1,4 @@
-package test;
+package graph;
 
 import java.nio.charset.StandardCharsets;
 import java.util.Date;
