+++ src/server/HTTPServer.java
@@ -1,4 +1,6 @@
-package test;
+package server;
+
+import servlets.Servlet;
 
 public interface HTTPServer extends Runnable {
