+++ src/server/MyHTTPServer.java
@@ -1,4 +1,6 @@
-package test;
+package server;
+
+import servlets.Servlet;
 
 import java.io.BufferedReader;
 import java.io.IOException;
