+++ src/server/RequestParser.java
@@ -1,4 +1,4 @@
-package test;
+package server;
 
 import java.io.BufferedReader;
 import java.io.File;
